requirements.txt
# Core dependencies
tkinter-tooltip==2.1.1
Pillow>=10.0.0
requests>=2.31.0
flask>=2.3.0
flask-restful>=0.3.10
flask-cors>=4.0.0

# Database
sqlalchemy>=2.0.0
sqlite3-fts4
alembic>=1.11.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
structlog>=23.1.0
rich>=13.4.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0
jwt>=1.3.1

# Machine Learning and Analytics
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# File handling and utilities
watchdog>=3.0.0
pathlib2>=2.3.7
magic>=0.4.27
chardet>=5.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0

# API and networking
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0

# Threading and async
asyncio
concurrent.futures
threading

# Cross-platform compatibility
psutil>=5.9.0
platform